requirements.txt
python-telegram-bot~=13.10
firebase_admin~=5.2.0

